% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_root.R
\name{get_root}
\alias{get_root}
\title{Collapse parts-of-speech differences}
\usage{
get_root(x)
}
\arguments{
\item{x}{A character vector of cleaned tokens}
}
\value{
A character vector of word-roots
}
\description{
\code{get_root()} attempts to retrieve the "root" of adjectives, descriptive
verbs, and plurals by removing common adjective and plural endings from the
words in \code{x}. This is best used to combine multiple forms of words which
exist in your dataset as both adjectives and nouns, to avoid splitting your
variables. It's a bit like using an internal dictionary, and it \emph{does} use a
dictionary (hunspell) to try and navigate the fact that some roots ending in
"e" (e.g., "cheese") lose their "e" when a suffix is added (e.g., "cheesy").
}
\details{
Please check \code{get_root()}'s performance with some selected words from your
data to ensure that they're combining and outputting correctly for your
intended use. This function prioritizes collapsing redundant word forms where
possible over returning pretty, perfect English--some words are prone to
returning fragments like "hon" for honey. You can always choose
(systematically or at random) a specific extension of the word-root that
appears in your data as the "display" form of the root.
}
\examples{
x <- c("tart", "tartness", "tarts", "fruit", "fruity", "fruitiness", "fruits",
"whiskey", "whisky", "nuts", "nut", "nutty", "rotten", "rotted", "rot",
"peppery", "plummy", "cheese", "cheesy")
get_root(x)
}
