% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{units_volume}
\alias{units_volume}
\title{Conversions for Units of Volume}
\format{
\subsection{\code{units_volume}}{

A named numeric vector with 47 entries, where the names are different
abbreviations for common units of volume and their conversion factors to SI
liters. Only US and SI units are currently included.
}
}
\source{
\url{https://www.nist.gov/pml/special-publication-811/nist-guide-si-appendix-b-conversion-factors/nist-guide-si-appendix-b9#VOLUME}
}
\usage{
units_volume
}
\description{
Common units of volume expressed in liters.
}
\keyword{datasets}
